pyproject.toml
[project]
name = "athan"
version = "0.1.0"
description = "Discord bot for accurate prayer times with voice Adhan support"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord.py==2.4.0",
    "aiohttp>=3.10.0",
    "aiosqlite==0.20.0",
    "python-dotenv==1.0.1",
    "tzdata==2024.1",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "PyNaCl>=1.5.0",
    "audioop-lts>=0.2.1",
]

[project.optional-dependencies]
dev = [
    "ruff==0.6.9",
    "black==24.10.0",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "freezegun==1.5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "PL"]
ignore = ["PLR0913", "PLR2004"]

[tool.ruff.lint.isort]
known-first-party = ["athan"]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["src"]

